Fix CHERI256 128-bit addition in decompress_256cap.
To compute the value of the capability register top, decompress_256cap performs 128-bit addition of
base and length. This is unintentionally performing a 64-bit addition of base and length and then
the result is promoted to a 128-bit integer, therefore any carry from the addition is not retained.
This causes clc/csc sequences on non-capability data to sometime zero-out the length portion of the
capability.
The 128-bit addition in decompress_256cap is fixed to work as expected.