Remove cap_register_t and add a dummy cc256_cap_t in its place
A QEMU that wants to use anything other than cc128_cap_t needs to define
HAVE_CAP_REGISTER_T, but due to the use of it in this header it also
needs to have defined cap_register_t before including this header,
which creates a circular dependency if QEMU wants to typedef it to the
underlying cc64_cap_t/cc128_cap_t itself as appropriate (#define would
work, but is ugly). However, in a world where we have both 64-bit and
128-bit compression implemented, having cap_register_t be defined by
cheri_compressed_cap.h doesn't make any sense, as both could easily be
wanted. Thus, simply delete it, and add a dummy cc256_cap_t that maps to
cc128_cap_t for the CHERI-256 functions in order to preserve the
existing behaviour, even if that perhaps doesn't make complete sense
(though the real solution is to continue the CHERI-256 purge).