Fix Morello reset capability encoding and de-duplicate
Morello uses an all-ones exponent for NULL but matches CHERI in using
the maximum encodable exponent for the almighty reset capability.
Previously we used the former encoding for the almighty reset
capability, which decodes to the same bounds but has a different
representation.

This whole set of definitions is also unnecessarily duplicated, so,
whilst here, merge them into one common definition.

Note that CCn_NULL_TOP already exists, and is the full decoded top, so
we use the name CCn_NULL_T instead and update CCn_RESET_TOP to match.
These are better names anyway, as they're values for T, the unshifted
value that is used to compute top, not top itself.