Move CapRegState into a new cap_register_t field
This allows us to simply copy cap_register_t values and not have to worry
about updating the external capreg_state array.
Longer-term it would be good to have cheri-compressed-cap transparently
handle the bounds decoded/raw bits states, but for now it is a lot easier
to add a cr_extra field to retain the existing meaning of the CREG_*
constants. This also means we don't have to adjust the Morello TCG code
that uses these constants.