Fix random_inputs_test for Morello
We have to XOR the pesbt field with the Morello NULL mask for the
compression library but we don't do so for the sail. This resulted in
an inconsistency when when converting to/from cc128m_cap_t.
Additionally, the existing XOR used the wrong XOR mask (it was using the
RISC-V one instead of the Morello one).