Fix incorrect behaviour in is_representable_with_addr
This was found by the fuzzer and was cuased by incorrectly masking of
the cursor bits inside is_representable_with_addr. We were triggering
the `"Input capability bounds not representable? They should have been rounded before!"`
assertion with the new test case. Instead of trying to re-create the
capability from scratch (which might not work with non-canonical bounds
encoding), create a new capability with the same pesbt and different
cursor and check if it decodes to the same value.