Change is_representable_with_addr API to select type of check
We shouldn't be relying on a magic _CC_USE_FAST_REP_CHECK macro to
change the behaviour of the bounds check. Instead just pass it as an
explicit argument. Also expose a new precise_is_representable_new_addr
function that complements the existing fast_is_representable_new_addr.
The important difference is whether the checks are exact vs approximate
not whether they are slow or fast (although I kept the fast name for the
approximate check since it matches the sail).