Avoid asserting on capabilities with bounds > MAX_TOP
This should never be possible (at least for RISC-V, Morello is somewhat
different due to SetTag). But if it does happen we should not assert.
Detagging would be one solution here, but for now we follow the behaviour
of the RISC-V and Morello models and just keep the tag as long as the
bounds change is monotonic.