Initial version of new 64r format
This is an initial draft of the 64r capability format for 64-bit
capabilities as defined in the RISC-V CHERI specification.

Note that this does not handle the decoding of raw permissions to the
architectural ones.
Future commits will add a new encode and decode function, but for now
just set CC64R_PERMS_ALL to the infinite cap value.

Co-authored-by: Alex Richardson <Alexander.Richardson@cl.cam.ac.uk>