Add a ENCODED_INFINITE_PERMS macro
This is the first step towards avoiding the UPERMS/HWPERMS fields
directly in the common code and makes the RV32 encoding of permissions
simpler. It will also allow removing the morello hacks since the uperms
there are in the middle of the permissions field.

Note: this needs to be a function macro since otherwise using
_CC_N(ENCODED_INFINITE_PERMS) gives us an error due to insufficient
pre-processor expansion.