Avoid using a "fake" XOR mask for Morello
Instead follow what the sail model does and invert the exponent and IE
values when reading them as part of the decode. This will also be useful
for supporting the RISC-V standard encoding which also avoids the XOR
on load and instead deals with the inversion during decode.
Introduce new _CC_N({ENCODE,EXTRACT}_{IE,EXPONENT}) macros to deal with
the differences in formats instead of reading the raw field values.

This also includes various Morello test changes since the input PESBT
values now need to be passed without the previous mask applied.