Add set_permission helper functions
This encodes the architectural permissions in the capability, returning
true/false depending on whether it was possible to encode the requested
set of permissions. Currently there is no legalization of the input
permissions (e.g. for RISC-V ASR cannot be set without X) and we just
check whether it is encodable.

Co-authored-by: Martin Kaiser <martin.kaiser@codasip.com>