Add lvbits parameter for infinite capability function
Add a new lvbits parameter to the inifinite capability function. Only the
caller knows if Zcherilevels are supported and if so, how many levels are
used. For now, we limit lvbits to 1. This gives us two levels (local, global).

We do not provide the make_max_perms_cap_ext API for Morello since this
avoids the need to define a fake _cc_mode for Morello.

Co-authored-by: Alexander Richardson <alexrichardson@google.com>