poison-cap: Introduce poison capabilities to RISC-V V9 cheri.
Introduce a new capability POISON field to mark poison capabilities.
This consumes one of the reserved bits in the V9 CHERI encoding.
The POISON bit is initialized as 0.

Introduce a new capability hardware permission bit PERM_POISON.
This is used to authorise privileged poison capability operations.
This consumes one of the reserved bits in the V9 CHERI encoding,
extending the HWPERMS by one bit, to 12 bits in total.
The PERM_POISON bit is initialized as 1 (similar to other permissions).

Note that this patch consumes all reserved bits in the V9 encoding.
It may be possible to recover some of these bits by either reducing
the SW permission bits (and steal one for PERM_POISON), reducing
otype space or reserving an otype for POISON capabilities.

Co-authored-by: Yuecheng Wang <yw737@cl.cam.ac.uk>